rename `file_finder.py` -> `level_iterator.py`
