+++ README.md
@@ -117,7 +117,7 @@ Actions in `SafeLifeGame` do not typically result in any direct rewards (there i
 
 The `safelife.safelife_env.SafeLifeEnv` class wraps `SafeLifeGame` in an interface suitable for reinforcement learning agents (à la [OpenAI Gym](https://gym.openai.com/)). It implements `step()` and `reset()` functions. The former accepts an action (integers 0–8) and outputs an observation, reward, whether or not the episode completed, and a dictionary of extra information (see the code for more details); the latter starts a new episode and returns a new observation. Observations in `SafeLifeEnv` are not the same as board states in `SafeLifeGame`. Crucially, the observation is always centered on the agent (this respects the symmetry of the game and means that agents don't have to implement attention mechanisms), can be partial (the agent only sees a certain distance), and only displays the color of the goal cells rather than their full content. The reward function in `SafeLifeEnv` is just the difference in point values between the board before and after an action and time-step update.
 
-Each `SafeLifeEnv` instance is initiated with a `level_iterator` object which generates new `SafeLifeGame` instances whenever the environment reset. The level iterator can most easily be created via `file_finder.safelife_loader` which can either load benchmark levels or generate new ones, e.g. `safelife_loader("benchmarks/v1.0/append-still")` or `safelife_loader("random/append-still")`. However, any function which generates `SafeLifeGame` instances would be suitable, and a custom method may be necessary to do e.g. curriculum learning.
+Each `SafeLifeEnv` instance is initiated with a `level_iterator` object which generates new `SafeLifeGame` instances whenever the environment reset. The level iterator can most easily be created via `level_iterator.SafeLifeLevelIterator` which can either load benchmark levels or generate new ones, e.g. `SafeLifeLevelIterator("benchmarks/v1.0/append-still")` or `SafeLifeLevelIterator("random/append-still")`. However, any function which generates `SafeLifeGame` instances would be suitable, and a custom method may be necessary to do e.g. curriculum learning.
 
 A gym environment is registered for each of the following level types:
 - `append-still`
